暂时完成，接着进阶

+++ public/js/admin.js
@@ -1,21 +1,23 @@
 $(function () {
     $('.del').click(function (e) {
-        alert('你确定要删除吗？');
-        let target = $(e.target);
-        let id = target.data('id');
-        let tr = $('.item-id-' + id);
+        let returnVal = confirm('你确定要删除吗？');
+        if(returnVal) {
+            let target = $(e.target);
+            let id = target.data('id');
+            let tr = $('.item-id-' + id);
 
-        $.ajax({
-            type: 'DELETE',
-            url: '/admin/list?id=' + id
-        })
-        .done((results) => {
-            if(results.success === 1) {
-                if(tr) {
-                    tr.remove();
-                }
-            }
-        });
+            $.ajax({
+                type: 'DELETE',
+                url: '/admin/list?id=' + id
+            })
+                .done((results) => {
+                    if(results.success === 1) {
+                        if(tr) {
+                            tr.remove();
+                        }
+                    }
+                });
+        }
     });
 });
